CustomSSLCertificateVerifier を別ファイルに移動
